scheduling: hoist the mail guards shared by notice tasks
## Summary

Moves `EmailCapExceeded` and `NotificationsDisabled` out of
`tasks/expiration_notices.py` into `tasks/mail_guards.py`, re-exported
from their old home so no caller changes.

Both are safety properties rather than conveniences, and both are about to
have a second consumer:

- `NotificationsDisabled` is why a notice task fails red instead of
  recording every message `suppressed`, reporting `succeeded`, and going
  green having mailed nobody. The CronJob does not inherit `webapp.env`,
  so that is a live failure mode.
- `EmailCapExceeded` carries `task_detail`, which `runner._execute` merges
  into the ledger row — the audience and the cap land as structured data
  rather than as a substring of `repr(exc)`.

Copying either into the next task is where one of them quietly diverges.
They go in their own module rather than `scheduling/registry.py` because
the runner knows nothing about mail and a task that sends none should not
import them.

`tests/unit/test_task_expiration_notices.py` passes **unedited** — that is
the proof the move was pure.

## Test Results

`test_task_expiration_notices.py` + `test_task_ledger.py` — 104 passed.
The per-file AST gate still holds: `mail_guards.py` imports nothing.
